Remove PEOPLEDB
According to Hyder, the peopledb was an experimental attempt to track
attendance and analyze trends for brigades. It's not a thing anymore,
and it must have recently been cleaned up because the `run_update.py`
script crashes when trying to connect to it.

There is some more cleanup to do regarding Attendance records, but I
will leave that for now.