Add `logo_url` to Organization API response
This adds a new field to the Organization API response that contains a
URL to a logo for that organization. Currently this is only pulled from
the GitHub projects_list_url, which is not the be-all-end-all for a
couple reasons:

1- It's hard to tell if an organization has uploaded a logo or if github
generated them a default blocky logo.

2- Some organizations use a CSV file for their projects_list_url.

In the case of #1 we will currently use the placeholder logo. In the
case of #2, the logo_url will be null. This adds support for, but I have
not tested, a "logo_url" key/value in the brigade-information
organizations.json file also, for any brigades that wish to override
this with a custom image.