Update "upcoming_events" to return events that haven't ended yet
This updates the notion of "upcoming" and "past" event to return events
based on their _end_ time instead of their _start_ time.

We got feedback from users that they were expecting to be able to find
the meetup they were current at (or late to) on the Brigade profile
page. To quickly meet this user need, we can just update the end time. I
assume that we are the primary user of this API and thus provides no
backwards compatibility. (Happy to add an option to this endpoint to
restore the old behavior if anyone notices this change.)