Remove cffi dependency
This package results in a build error on Heroku.

We don't explicitly use this dependency ourselves, it was added "to
support secure connections" but I am not sure what that means. Let's
assume pip will install this if necessary, and if this causes a
regression we will see it in Sentry.